Resizable: Remove bad workaround for draggable+resizable
This adds a compound test page for draggable+resizable, which had no coverage
before. Using that page shows that there is no way to reproduce the behaviour
described in the original ticket that caused this workaround, since its not
possible to resize an element beyond the window boundaries. Therefore removing
the workaround, which is 6+ years old and has no test coverage, seems like the
sanest approach.

The other half of the workaround, checking for position:absolute, still makes
sense, as the updated comment indicates. Without the explicit position, the
relative positioned handles don't move along when resizing.

Fixes #6939